+++ src/main/java/com/example/Todolist/Task/TaskService.java
@@ -4,6 +4,9 @@ import com.example.Todolist.Board.Board;
 import com.example.Todolist.Board.BoardRepository;
 import com.example.Todolist.Exceptions.AlreadyDefined;
 import com.example.Todolist.Exceptions.NotFound;
+import com.example.Todolist.Exceptions.NotFoundUser;
+import com.example.Todolist.User.User;
+import com.example.Todolist.User.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -12,12 +15,18 @@ import org.springframework.stereotype.Service;
 public class TaskService {
     private final TaskRepository taskRepository;
     private final BoardRepository boardRepository;
-    public void save(Task task) throws AlreadyDefined, NotFound {
+    private final UserRepository userRepository;
+    public void save(Task task, String userid) throws AlreadyDefined, NotFound, NotFoundUser {
         Task oldTask = taskRepository.findByName(task.getName());
         Board board = boardRepository.findByUuid(task.getBoard_uuid());
-        if (oldTask == null && board != null) {
+        User user = userRepository.findUserById(userid);
+        if (oldTask == null && board != null && user != null) {
+            task.setUserid(userid);
             taskRepository.save(task);
         } else {
+            if (user == null) {
+                throw new NotFoundUser("Can't find User with this id");
+            }
             if (oldTask != null) {
             throw new AlreadyDefined("Task with this name already defined");
             }
@@ -26,23 +35,45 @@ public class TaskService {
             }
         }
     }
-    public void update(Task task, String uuid) throws NotFound {
+    public void update(Task task, String uuid, String userid) throws NotFound, NotFoundUser {
         Task oldTask = taskRepository.findByUuid(uuid);
+        User user = userRepository.findUserById(userid);
+        if (user == null) {
+            throw new NotFoundUser("Can't find User with this id");
+        }
         if (oldTask == null) {
             throw new NotFound("Task with this uuid does not exist");
         }
         oldTask.setName(task.getName());
         taskRepository.save(oldTask);
     }
-    public Task findByUuid(String uuid) throws NotFound {
-        return taskRepository.findById(uuid).orElseThrow(() -> new NotFound("Can't found task with this uuid"));
+    public Task findByUuid(String uuid, String userid) throws NotFound, NotFoundUser {
+        Task task = taskRepository.findByUuid(uuid);
+        User user = userRepository.findUserById(userid);
+        if (task != null && user != null) {
+            return task;
+        } else {
+            if (user == null) {
+                throw new NotFoundUser("Can't find User with this id");
+            }
+            if (task == null) {
+                throw new NotFound("Task with this uuid does not exist");
+            }
+            return null;
+        }
     }
-    public void delete(String uuid) throws NotFound {
+    public void delete(String uuid, String userid) throws NotFound, NotFoundUser {
         Task taskToDelete = taskRepository.findByUuid(uuid);
-        if(taskToDelete != null) {
+        User user = userRepository.findUserById(userid);
+        if(taskToDelete != null && user != null) {
             taskRepository.deleteById(uuid);
         } else {
-            throw new NotFound("Can't found task with this uuid");
+            if (user == null) {
+                throw new NotFoundUser("Can't find User with this id");
+            }
+            if (taskToDelete == null) {
+                throw new NotFound("Can't found task with this uuid");
+            }
         }
     }
 }
